EKS/Strands/Bedrock-Agentcore/requirements.txt
strands-agents
strands-agents-tools
bedrock-agentcore